Swap names of set and configure commands
