+++ lib/storage/commands/configure.rb
@@ -25,6 +25,7 @@
 # https://github.com/openflighthpc/flight-storage
 #==============================================================================
 require 'tty-prompt'
+require 'yaml'
 
 require_relative '../client_factory'
 require_relative '../command'
@@ -33,20 +34,33 @@ module Storage
   module Commands
     class Configure < Command
       def run
-        provider = prompt.select(
-          "Select your desired cloud provider:",
-          choices
-        )
+        klass = ClientFactory::PROVIDERS[Config.provider][:klass]
+        questions = to_questions(klass)
+        answers = prompt.collect do
+          questions.each do |question|
+            key(question[:key]).ask(question[:text], required: true)
+          end
+        end
 
-        Config.data.set(:provider, value: provider)
-        Config.save_data
+        if save_credentials(answers)
+          puts "Credentials saved to #{Config.credentials_dir}/"
+        end
       end
 
       private
 
-      def choices
-        ClientFactory::PROVIDERS.map do |k, v|
-          { v[:friendly_name] => k.to_s }
+      def save_credentials(credentials_hash)
+        File.write(Config.credentials_file, YAML.dump(credentials_hash))
+      end
+
+      def to_questions(klass)
+        klass.creds_schema.map do |k, v|
+          friendly = k.to_s.gsub('_', ' ').capitalize
+          text = "#{friendly}:"
+          {
+            key: k,
+            text: text
+          }
         end
       end
